pyproject.toml
[tool.poetry]
name = "py-aiger-jani"
version = "0.1.0"
description = "Python library for translating from a subset of Jani to AIGs."
repository = "https://github.com/mvcisback/py-aiger-jani"
authors = ["Marcell Vazquez-Chanlatte <mvc@linux.com>"]
readme="README.md"
license = "MIT"
packages = [
    { include = "aiger_jani" },
]

[tool.poetry.dependencies]
python = "^3.9"
py-aiger = "^6.1.14"
py-aiger-bv = "^4.5.3"
numpy = "^1.20.2"
attrs = "^20.3.0"
py-aiger-coins = {version = "^3.3.3", extras = ["=bdd", "=sat"]}
networkx = "^2.5.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
pytest-cov = "^2.11.1"
pytest-flake8 = "^1.0.7"
coverage = "^5.5"
codecov = "^2.1.11"
pdoc3 = "^0.9.2"
py-aiger-dfa = "^0.3.1"
mdd = "^0.3.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
